requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.4
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.14
django-rest-framework==0.1.0
djangorestframework==3.11.2
django-celery-results==1.2.1
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.10
mysqlclient==1.4.6
plaid-python==4.0.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0